V 1.0.3
complete web assembly source
add barcode scanner to web assembly
add local storage to web assembly
change view model and app structure

+++ composeApp/src/commonMain/kotlin/io/domil/store/view/SnackBar.kt
@@ -89,7 +89,7 @@ fun showLog(
         state.showSnackbar(
             data,
             action.toString(),
-            if (action == SnackBarActions.WARNING) SnackbarDuration.Indefinite else SnackbarDuration.Long
+            if (action == SnackBarActions.WARNING) SnackbarDuration.Indefinite else SnackbarDuration.Short
         )
     }
 }
